research_site/requirements.txt
django
django-storages
django-tinymce
boto3
numpy
pandas
plotly
gunicorn
pillow
whitenoise
psycopg2
